Update bulk operations to produce consistent response
* Return same DTO object
* Update response codes
