+++ src/main/java/com/czertainly/api/interfaces/core/web/ClientManagementController.java
@@ -5,7 +5,7 @@ import com.czertainly.api.exception.NotFoundException;
 import com.czertainly.api.exception.ValidationException;
 import com.czertainly.api.model.client.client.AddClientRequestDto;
 import com.czertainly.api.model.client.client.EditClientRequestDto;
-import com.czertainly.api.model.client.connector.ForceDeleteMessageDto;
+import com.czertainly.api.model.common.BulkActionMessageDto;
 import com.czertainly.api.model.client.raprofile.SimplifiedRaProfileDto;
 import com.czertainly.api.model.common.ErrorMessageDto;
 import com.czertainly.api.model.common.UuidDto;
@@ -93,7 +93,7 @@ public interface ClientManagementController {
     @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "Clients Deleted")})
     @RequestMapping(method = RequestMethod.DELETE, produces = {"application/json"})
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public ResponseEntity<List<ForceDeleteMessageDto>> bulkRemoveClient(@io.swagger.v3.oas.annotations.parameters.RequestBody(
+    public ResponseEntity<List<BulkActionMessageDto>> bulkRemoveClient(@io.swagger.v3.oas.annotations.parameters.RequestBody(
             description = "Client UUIDs", content = @Content(array = @ArraySchema(schema = @Schema(implementation = String.class)),
             examples={@ExampleObject(value="[\"c2f685d4-6a3e-11ec-90d6-0242ac120003\",\"b9b09548-a97c-4c6a-a06a-e4ee6fc2da98\"]")}))
                                      @RequestBody List<String> clientUuids) throws NotFoundException;
