+++ src/main/java/com/czertainly/api/interfaces/core/web/ConnectorController.java
@@ -178,18 +178,17 @@ public interface ConnectorController {
 			@ApiResponse(responseCode = "422", description = "Unprocessible Entity", content = @Content(array = @ArraySchema(schema = @Schema(implementation = String.class)),
 					examples = {@ExampleObject(value = "[\"Error Message 1\",\"Error Message 2\"]")}))})
 	@RequestMapping(method = RequestMethod.DELETE, produces = {"application/json"})
-	public List<ForceDeleteMessageDto> bulkRemoveConnector(@io.swagger.v3.oas.annotations.parameters.RequestBody(
+	public List<BulkActionMessageDto> bulkRemoveConnector(@io.swagger.v3.oas.annotations.parameters.RequestBody(
 			description = "Connector UUIDs", content = @Content(array = @ArraySchema(schema = @Schema(implementation = String.class)),
 			examples={@ExampleObject(value="[\"c2f685d4-6a3e-11ec-90d6-0242ac120003\",\"b9b09548-a97c-4c6a-a06a-e4ee6fc2da98\"]")}))
 															   @RequestBody List<String> uuids) throws NotFoundException, ValidationException, ConnectorException;
 
 	@Operation(summary = "Force Delete multiple Connectors")
-	@ApiResponses(value = { @ApiResponse(responseCode = "204", description = "Connectors deleted"),
+	@ApiResponses(value = { @ApiResponse(responseCode = "200", description = "Connectors deleted"),
 			@ApiResponse(responseCode = "422", description = "Unprocessible Entity", content = @Content(array = @ArraySchema(schema = @Schema(implementation = String.class)),
 					examples = {@ExampleObject(value = "[\"Error Message 1\",\"Error Message 2\"]")}))})
 	@RequestMapping(path = "/force", method = RequestMethod.DELETE, produces = {"application/json"})
-	@ResponseStatus(HttpStatus.NO_CONTENT)
-	public void bulkForceRemoveConnector(@io.swagger.v3.oas.annotations.parameters.RequestBody(
+	public List<BulkActionMessageDto> bulkForceRemoveConnector(@io.swagger.v3.oas.annotations.parameters.RequestBody(
 			description = "Connector UUIDs", content = @Content(array = @ArraySchema(schema = @Schema(implementation = String.class)),
 			examples={@ExampleObject(value="[\"c2f685d4-6a3e-11ec-90d6-0242ac120003\",\"b9b09548-a97c-4c6a-a06a-e4ee6fc2da98\"]")}))
 											 @RequestBody List<String> uuids) throws NotFoundException, ValidationException;
