+++ src/main/java/com/czertainly/api/interfaces/core/web/CredentialController.java
@@ -4,7 +4,7 @@ import com.czertainly.api.exception.AlreadyExistException;
 import com.czertainly.api.exception.ConnectorException;
 import com.czertainly.api.exception.NotFoundException;
 import com.czertainly.api.exception.ValidationException;
-import com.czertainly.api.model.client.connector.ForceDeleteMessageDto;
+import com.czertainly.api.model.common.BulkActionMessageDto;
 import com.czertainly.api.model.client.credential.CredentialRequestDto;
 import com.czertainly.api.model.client.credential.CredentialUpdateRequestDto;
 import com.czertainly.api.model.common.ErrorMessageDto;
@@ -111,7 +111,8 @@ public interface CredentialController {
             @ApiResponse(responseCode = "422", description = "Unprocessible Entity", content = @Content(array = @ArraySchema(schema = @Schema(implementation = String.class)),
                     examples = {@ExampleObject(value = "[\"Error Message 1\",\"Error Message 2\"]")}))})
     @RequestMapping(method = RequestMethod.DELETE, produces = {"application/json"})
-    public List<ForceDeleteMessageDto> bulkRemoveCredential(@io.swagger.v3.oas.annotations.parameters.RequestBody(
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void bulkRemoveCredential(@io.swagger.v3.oas.annotations.parameters.RequestBody(
             description = "Credential UUIDs", content = @Content(array = @ArraySchema(schema = @Schema(implementation = String.class)),
             examples = {@ExampleObject(value = "[\"c2f685d4-6a3e-11ec-90d6-0242ac120003\",\"b9b09548-a97c-4c6a-a06a-e4ee6fc2da98\"]")}))
                                                             @RequestBody List<String> uuids) throws NotFoundException, ValidationException;
